new commit

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,14 +8,14 @@
 int main(void)
 {
 	int i;
-	long long a, b, sum;
+	unsigned long a, b, sum;
 
 	a = 0;
 	b = 1;
 	for (i = 0; i < 50; i++)
 	{
 		sum = a + b;
-		printf("%lld", sum);
+		printf("%lu", sum);
 		if (i != 49)
 		{
 			printf(",");
